pyproject.toml
[project]
name = "michael-retirement-travel-planner"
version = "1.0.0"
description = "LangChain-powered AI agent for multi-city retirement travel planning"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.2.5",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.81",
    "langchain-core>=0.3.81",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.hatch.build.targets.wheel]
include = [
    "*.py",
    "tools/**/*.py"
]
